refactor: altera o namespace pois foram criadas subpastas para melhor organização
