+++ Semana-5/Comex.Testes/ClienteTestes.cs
@@ -1,4 +1,5 @@
-﻿using FluentAssertions;
+﻿using Comex.Models;
+using FluentAssertions;
 
 namespace Comex.Testes
 {
