+++ Semana-5/Comex.Testes/ProdutoTestes.cs
@@ -1,4 +1,5 @@
-﻿using FluentAssertions;
+﻿using Comex.Models;
+using FluentAssertions;
 
 namespace Comex.Testes
 {
