+++ Semana-5/Comex/Models/Estoque.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Comex
+namespace Comex.Models
 {
     public class Estoque
     {
