+++ Semana-5/Comex/Models/ProdutoIsento.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Comex
+namespace Comex.Models
 {
     public class ProdutoIsento : Produto
     {
