+++ Semana-5/Comex/TestaModels/TestaProduto.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Comex.Models;
 
-namespace Comex
+namespace Comex.TestaModels
 {
     public class TestaProduto
     {
